Finally bit the bullet and renamed the classes.
CJSONDataSerializer is now CJSONSerializer.
The old CJSONSerializer has now gone away.

Why? CJSONSerializer serialized to strings. You really should not be doing that anyways. And if you need to - just serialize to an NSData and convert to a string.

If you used CJSONSerializer and expected to get an NSString you'll now get an NSData. This could be bad for you.