Moved error codes to typedef'd enums
Moved various hardcoded error codes into enums that are defined near their corresponding error domain constants. The intent is in part to enable library users to more easily grasp the meaning of these codes as they encounter them and to simplify doing something with them. (e.g. displaying custom error messages on a per-error basis.)

These codes have been transcribed verbatim from the original source and named based on the existing error messages. The mapping of codes to the actual detection of error states in CJSONScanner had a few possible anomalies that are still reflected here. Resolution of these anomalies (as appropriate) is intended for the next commit.